requirements.txt
flask
mysql-connector-python
cryptography
schedule
scikit-learn
pandas
google-cloud-storage
python-dotenv
werkzeug
